19-10-23 Crud Book

+++ 23-24/3A43/src/Controller/AuthorController.php
@@ -79,8 +79,14 @@ class AuthorController extends AbstractController
     {
         $author= $repository->find($id);
         $em= $managerRegistry->getManager();
-        $em->remove($author);
-        $em->flush();
+        if($author->getNbBooks()==0){
+            $em->remove($author);
+            $em->flush();
+        }
+        else{
+            return new Response("Error");
+        }
+
         return $this->redirectToRoute("authors_list");
 
     }
